renamed all project name instances to emuLadder
