+++ angular/README.md
@@ -1,4 +1,4 @@
-# Capstone
+# emuLadder
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 8.1.2.
