+++ angular/angular.json
@@ -3,7 +3,7 @@
   "version": 1,
   "newProjectRoot": "projects",
   "projects": {
-    "capstone": {
+    "emuLadder": {
       "projectType": "application",
       "schematics": {},
       "root": "",
