+++ angular/src/app/app.component.spec.ts
@@ -20,16 +20,16 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'capstone'`, () => {
+  it(`should have as title 'emuLadder'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
-    expect(app.title).toEqual('capstone');
+    expect(app.title).toEqual('emuLadder');
   });
 
   it('should render title in a h1 tag', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector('h1').textContent).toContain('Welcome to capstone!');
+    expect(compiled.querySelector('h1').textContent).toContain('Welcome to emuLadder!');
   });
 });
