+++ angular/src/favicon.ico
