+++ java/src/main/java/com/uncg/emuLadder/Application.java
@@ -1,4 +1,4 @@
-package com.uncg.capstone;
+package com.uncg.emuLadder;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
